refactor: split methods for Tools class into separate mixin classes.
The Tools class initially was more than 1000 lines long, which is too 
bulky to maintain and read effectively. I broke it out into three 
separate 'mixin' classes. They have no __init__ method because they just 
serve to provide methods to the Tools class when they are inherited.
